ai_module.py
import requests
import spacy
import nltk
from nltk.wsd import lesk
from nltk import word_tokenize
from transformers import pipeline
from textblob import TextBlob
from deep_translator import GoogleTranslator
import numpy as np
from sklearn.feature_extraction.text import TfidfVectorizer
from sklearn.cluster import KMeans
from gtts import gTTS

import matplotlib.pyplot as plt
from PIL import Image
import os
import uuid

class AdvancedAI:
    def __init__(self):
        # Load necessary models
        self.nlp = spacy.load('en_core_web_sm')
        self.generator = pipeline('text-generation', model='gpt2', tokenizer='gpt2')
        # Initialize emotional state
        self.emotional_state = {'mood': 'Neutral', 'intensity': 0.0}
        # Emotional thresholds
        self.emotional_thresholds = {
            'Very Positive': 0.75,
            'Positive': 0.25,
            'Neutral': 0.0,
            'Negative': -0.25,
            'Very Negative': -0.75
        }
        # Self-talk phrases
        self.self_talk_phrases = [
            "Stay calm.",
            "Let's think this through.",
            "Maintain composure.",
            "Focus on the positive.",
            "Take a deep breath."
        ]
        # Initialize other attributes for various intelligences
        self.network = {}
        self.relationships = {}
        self.knowledge_base = {}
        # Initialize conversation history
        self.history = []
        print("AdvancedAI initialized with all functionalities.")

    def handle_user_input(self, text, consent, mode):
        """
        Main method to handle user input and generate response.
        """
        # Integrate multiple forms of intelligence
        self.demonstrate_intelligences(text, mode)

        if mode == 'emotion':
            # Analyze emotion
            emotion_data = self.analyze_emotion(text)
            self.manage_emotions()
            emotion_text = f"Detected emotion: {emotion_data['emotion']}."
        else:
            emotion_data = {'emotion': 'Not Analyzed', 'polarity': 0}
            emotion_text = ""

        # Process text
        processed_text = self.process_text(text)

        # Generate creative text if prompted
        if 'create' in text.lower() or 'generate' in text.lower():
            creative_prompt = text
            creative_text = self.generate_creative_text(creative_prompt)
            response_text = creative_text
        else:
            # Autonomous problem-solving
            solution = self.solve_problem(text)
            response_text = solution

        if mode == 'emotion' and emotion_text:
            response_text += f" {emotion_text}"

        # Update history based on user consent
        if consent:
            self.history.append({
                'user_input': text,
                'response': response_text,
                'emotion': emotion_data['emotion']
            })
            # Data can be used for AI learning and development
            self.learn_from_data(text)
        else:
            # Store data in user's personal history only
            pass  # Implement storing in user's personal history if needed

        # Prepare response
        response = {
            'response': response_text,
            'emotion': emotion_data['emotion'],
            'history': self.history
        }
        return response

    def get_latest_information(self, topic):
        """
        Fetch the latest information on a topic using the Wikipedia API.
        """
        topic_formatted = topic.replace(' ', '_')
        url = f'https://en.wikipedia.org/api/rest_v1/page/summary/{topic_formatted}'
        response = requests.get(url)
        if response.status_code == 200:
            data = response.json()
            return data.get('extract', 'No information found.')
        else:
            return "Topic not found."

    def process_text(self, text):
        """
        Process text to understand context, entities, and syntax.
        """
        doc = self.nlp(text)
        entities = [(ent.text, ent.label_) for ent in doc.ents]
        tokens = [(token.text, token.pos_, token.dep_) for token in doc]
        return {'entities': entities, 'tokens': tokens}

    def detect_sarcasm(self, text):
        """
        Detect sarcasm using placeholder logic.
        """
        sarcastic_keywords = ["yeah right", "sure", "as if", "totally", "great", "wonderful"]
        if any(phrase in text.lower() for phrase in sarcastic_keywords):
            return "Sarcasm may be present."
        else:
            return "No sarcasm detected."

    def disambiguate_word(self, sentence, word):
        """
        Use Word Sense Disambiguation to find the meaning of a word in context.
        """
        sense = lesk(word_tokenize(sentence), word)
        if sense:
            return sense.definition()
        else:
            return "No definition found."

    def generate_creative_text(self, prompt):
        """
        Generate creative text based on a prompt.
        """
        generated = self.generator(prompt, max_length=50, num_return_sequences=1)
        return generated[0]['generated_text']

    def analyze_emotion(self, text):
        """
        Analyze the emotion of the input text and update internal emotional state.
        """
        blob = TextBlob(text)
        polarity = blob.sentiment.polarity
        subjectivity = blob.sentiment.subjectivity

        # Determine emotion category
        if polarity >= self.emotional_thresholds['Very Positive']:
            emotion = 'Very Positive'
        elif self.emotional_thresholds['Positive'] <= polarity < self.emotional_thresholds['Very Positive']:
            emotion = 'Positive'
        elif self.emotional_thresholds['Negative'] < polarity < self.emotional_thresholds['Positive']:
            emotion = 'Neutral'
        elif self.emotional_thresholds['Very Negative'] <= polarity <= self.emotional_thresholds['Negative']:
            emotion = 'Negative'
        else:
            emotion = 'Very Negative'

        # Update internal emotional state
        self.emotional_state['mood'] = emotion
        self.emotional_state['intensity'] = polarity

        return {'emotion': emotion, 'polarity': polarity, 'subjectivity': subjectivity}

    def verbalize_emotion(self, emotion_data):
        """
        Verbally express the emotion analysis using text-to-speech.
        """
        emotion_text = f"The emotion detected is {emotion_data['emotion']} with a polarity of {emotion_data['polarity']:.2f}."
        tts = gTTS(text=emotion_text, lang='en')
        filename = f"emotion_{uuid.uuid4()}.mp3"
        tts.save(filename)
        os.remove(filename)

    def visualize_emotion(self, emotion_data):
        """
        Visually represent the emotion analysis using an image.
        """
        emotion = emotion_data['emotion']
        image_path = os.path.join('emotion_images', f"{emotion.lower().replace(' ', '_')}.png")

        if os.path.exists(image_path):
            img = Image.open(image_path)
            plt.imshow(img)
            plt.axis('off')
            plt.show()
        else:
            plt.figure(figsize=(6, 3))
            plt.text(0.5, 0.5, emotion, fontsize=24, ha='center', va='center')
            plt.axis('off')
            plt.show()

    def manage_emotions(self):
        """
        Manage internal emotional state appropriately.
        """
        mood = self.emotional_state['mood']
        intensity = self.emotional_state['intensity']

        if mood in ['Negative', 'Very Negative']:
            print("AI Self-Talk: Engaging in self-regulation strategies.")
            self.self_talk()
            # Adjust emotional state after self-talk
            self.emotional_state['mood'] = 'Neutral'
            self.emotional_state['intensity'] = 0.0
        else:
            print("AI is maintaining a positive or neutral emotional state.")

    def self_talk(self):
        """
        Simulate productive self-talk to prevent emotional hijacking.
        """
        for phrase in self.self_talk_phrases:
            print(f"AI Self-Talk: {phrase}")

    def demonstrate_intelligences(self, text, mode):
        """
        Demonstrate the integration of various forms of intelligence.
        """
        # Intellectual Intelligence (IQ)
        processed = self.process_text(text)
        # Emotional Intelligence (if mode == 'emotion')
        if mode == 'emotion':
            emotion_data = self.analyze_emotion(text)
            self.manage_emotions()
        # Social Intelligence
        # Additional analysis for other intelligences
        # Placeholder for detailed implementation

    def solve_problem(self, query):
        """
        Attempts to solve problems across various industries based on the query.
        """
        # Identify industry based on keywords
        industries = {
            'finance': ['finance', 'investment', 'bank', 'stock'],
            'healthcare': ['health', 'medicine', 'doctor', 'hospital'],
            'technology': ['technology', 'computer', 'software', 'hardware'],
            # Add more industries as needed
        }
        for industry, keywords in industries.items():
            if any(keyword in query.lower() for keyword in keywords):
                solution = f"As an AI specializing in {industry}, I suggest the following solution..."
                # Implement industry-specific logic here
                return solution
        # Default response
        return "I'm processing your query to find the best possible solution."

    def learn_from_data(self, text):
        """
        Simulate learning from user data.
        """
        # Placeholder for learning algorithm
        pass
