requirements.txt
flask
flask_login
flask_wtf
pyotp
qrcode
pillow
passlib
requests
nltk
transformers
textblob
torch
deep_translator
numpy
scikit-learn
gTTS
flask_sqlalchemy
spacy==3.6.1
matplotlib
pycryptodome==3.15.0
